selective disclosure tests
